combine modules to avoid circular dependency
