+++ bot/kodiak/test_queries.py
@@ -407,9 +407,7 @@ async def test_get_subscription_missing_blocker(
             b"subscription_blocker": b"",
         }
     )
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -431,9 +429,7 @@ async def test_get_subscription_missing_blocker_and_data(
             b"data": b"",
         }
     )
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -450,9 +446,7 @@ async def test_get_subscription_missing_blocker_fully(
     Redis. We should handle this case by returning an empty subscription.
     """
     fake_redis = create_fake_redis_reply({})
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res is None
@@ -473,9 +467,7 @@ async def test_get_subscription_seats_exceeded(
             b"data": b'{"kind":"seats_exceeded", "allowed_user_ids": [5234234]}',
         }
     )
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -498,9 +490,7 @@ async def test_get_subscription_seats_exceeded_no_seats(
             b"data": b'{"kind":"seats_exceeded", "allowed_user_ids": []}',
         }
     )
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -524,9 +514,7 @@ async def test_get_subscription_seats_exceeded_missing_data(
         }
     )
 
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -550,9 +538,7 @@ async def test_get_subscription_seats_exceeded_invalid_data(
         }
     )
 
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -577,9 +563,7 @@ async def test_get_subscription_seats_exceeded_invalid_kind(
         }
     )
 
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
@@ -602,9 +586,7 @@ async def test_get_subscription_unknown_blocker(
         }
     )
 
-    mocker.patch(
-        "kodiak.event_handlers.get_redis", return_value=wrap_future(fake_redis)
-    )
+    mocker.patch("kodiak.queue.get_redis", return_value=wrap_future(fake_redis))
     async with api_client as api_client:
         res = await api_client.get_subscription()
     assert res == Subscription(
